feat: extract AppDetailsCDModal component
